Remove unnecessary class
