+++ scm-webapp/src/main/java/sonia/scm/importexport/FullScmRepositoryImporter.java
@@ -24,10 +24,14 @@
 
 package sonia.scm.importexport;
 
+import org.apache.commons.compress.archivers.tar.TarArchiveEntry;
 import org.apache.commons.compress.archivers.tar.TarArchiveInputStream;
 import org.apache.commons.compress.compressors.gzip.GzipCompressorInputStream;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import sonia.scm.ContextEntry;
 import sonia.scm.repository.Repository;
+import sonia.scm.repository.RepositoryManager;
 import sonia.scm.repository.api.ImportFailedException;
 
 import javax.inject.Inject;
@@ -35,13 +39,24 @@ import java.io.BufferedInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 
+import static java.util.Arrays.stream;
+
 public class FullScmRepositoryImporter {
 
-  private final FullScmRepositoryImporterProcess process;
+  private static final Logger LOG = LoggerFactory.getLogger(FullScmRepositoryImporter.class);
+
+  private final ImportStep[] importSteps;
+  private final RepositoryManager repositoryManager;
 
   @Inject
-  public FullScmRepositoryImporter(FullScmRepositoryImporterProcess process) {
-    this.process = process;
+  public FullScmRepositoryImporter(EnvironmentCheckStep environmentCheckStep,
+                                   MetadataImportStep metadataImportStep,
+                                   StoreImportStep storeImportStep,
+                                   RepositoryImportStep repositoryImportStep,
+                                   RepositoryManager repositoryManager
+  ) {
+    this.repositoryManager = repositoryManager;
+    importSteps = new ImportStep[]{environmentCheckStep, metadataImportStep, storeImportStep, repositoryImportStep};
   }
 
   public Repository importFromStream(Repository repository, InputStream inputStream) {
@@ -52,7 +67,7 @@ public class FullScmRepositoryImporter {
           GzipCompressorInputStream gcis = new GzipCompressorInputStream(bif);
           TarArchiveInputStream tais = new TarArchiveInputStream(gcis)
         ) {
-          return process.run(repository, tais);
+          return run(repository, tais);
         }
       } else {
         throw new ImportFailedException(
@@ -67,5 +82,37 @@ public class FullScmRepositoryImporter {
         e
       );
     }
+  }
+
+  private Repository run(Repository repository, TarArchiveInputStream tais) throws IOException {
+    ImportState state = new ImportState(repositoryManager.create(repository));
+    try {
+      TarArchiveEntry tarArchiveEntry;
+      while ((tarArchiveEntry = tais.getNextTarEntry()) != null) {
+        LOG.trace("Trying to handle tar entry '{}'", tarArchiveEntry.getName());
+        handle(tais, state, tarArchiveEntry);
+      }
+      stream(importSteps).forEach(step -> step.finish(state));
+      return state.getRepository();
+    } finally {
+      stream(importSteps)
+        .forEach(step -> step.cleanup(state));
+      if (!state.success()) {
+        // Delete the repository if any error occurs during the import
+        repositoryManager.delete(state.getRepository());
+      }
+    }
+  }
+
+  private void handle(TarArchiveInputStream tais, ImportState state, TarArchiveEntry currentEntry) {
+    for (ImportStep step : importSteps) {
+      if (step.handle(currentEntry, state, tais)) {
+        return;
+      }
+    }
+    throw new ImportFailedException(
+      ContextEntry.ContextBuilder.entity(state.getRepository()).build(),
+      "Invalid import format. Unknown file in tar: " + currentEntry.getName()
+    );
   }
 }
